+++ src/com/GoldenMine/effects/EffectFadeOut.java
@@ -17,7 +17,7 @@ public class EffectFadeOut implements IEffect {
     }
 
     @Override
-    public void editImage(BufferedImage original, BufferedImage changed, Graphics changeGraphics, int percent) {
+    public void editImage(BufferedImage original, BufferedImage changed, Graphics changeGraphics, double percent) {
         int x = changed.getWidth();
         int y = changed.getHeight();
